% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_relationship.R
\name{add_relationship}
\alias{add_relationship}
\title{Add a relationship to an entityset}
\usage{
add_relationship(entityset, set1, set2, idx)
}
\arguments{
\item{entityset}{The entityset to modify.}

\item{set1}{The name of the first entity to link.}

\item{set2}{The name of the second entity to link.}

\item{idx}{The variable  `set1` and `set2` have in common.}
}
\value{
A modified entityset.
}
\description{
Add a relationship to an entityset.
}
\examples{
library(magrittr)
options(stringsAsFactors = TRUE)
set_1 <- data.frame(key = 1:100, value = sample(letters, 100, TRUE))
set_2 <- data.frame(key = 1:100, value = sample(LETTERS, 100, TRUE))
# Common variable: `key`

as_entityset(set_1, index = "key", entity_id = "set_1", id = "demo") \%>\%
  add_entity(entity_id = "set_2", df = set_2, index = "key") \%>\%
  add_relationship(set1 = "set_1", set2 = "set_2", idx = "key")
}
