% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as_entityset.R
\name{as_entityset}
\alias{as_entityset}
\title{Create entityset and entity from data frame.}
\usage{
as_entityset(.data, id = "entityset", index = NA, time_index = NULL,
  entity_id = "df1", ...)
}
\arguments{
\item{.data}{The `data.frame` to be added as an entity to entityset.}

\item{id}{The id of this entityset.}

\item{index}{Name of id column in the dataframe.}

\item{time_index}{Name of the time column in the dataframe.}

\item{entity_id}{An identifier for this entity.}

\item{...}{Additional variables passed to `add_entity`.}
}
\value{
A modified entityset.
}
\description{
Create an entityset with a selected `data.frame` as an entity.
}
\examples{
\donttest{
as_entityset(cars, index = "row_number")
}
}
