% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_feature_matrix.R
\name{calculate_feature_matrix}
\alias{calculate_feature_matrix}
\title{Calculate feature matrix}
\usage{
calculate_feature_matrix(entityset, features, ...)
}
\arguments{
\item{entityset}{The entityset on which to create features.}

\item{features}{The features to create based on previous runs of `dfs`.}

\item{...}{Additional parameters passed to `featuretoools.calculate_feature_matrix`.}
}
\value{
A feature matrix
}
\description{
This function is used to create a feature matrix based on a custom list of features (usually created from `save_features`).
}
