% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_entityset.R
\name{create_entityset}
\alias{create_entityset}
\title{Create entityset}
\usage{
create_entityset(id)
}
\arguments{
\item{id}{The id of this entityset.}
}
\value{
An entityset.
}
\description{
Create a blank entityset. A shortcut for `featuretools'` `EntitySet`.
}
\examples{
\donttest{
create_entityset(id = "my_entityset")
}
}
