% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_features.R
\name{extract_features}
\alias{extract_features}
\title{Extract features}
\usage{
extract_features(.data)
}
\arguments{
\item{.data}{The featuretools-object returned from `dfs`.}
}
\value{
All features created during dfs, as a tibble.
}
\description{
This function is used to extract all features created from `dfs(...)`.
}
\examples{
library(magrittr)
options(stringsAsFactors = T)
set_1 <- data.frame(key = 1:100, value = sample(letters, 100, T))
set_2 <- data.frame(key = 1:100, value = sample(LETTERS, 100, T))
# Common variable: `key`

as_entityset(set_1, index = "key", entity_id = "set_1", id = "demo") \%>\%
  add_entity(entity_id = "set_2", df = set_2, index = "key") \%>\%
  add_relationship(set1 = "set_1", set2 = "set_2", idx = "key") \%>\%
  dfs(target_entity = "set_1", trans_primitives = c("and", "divide")) \%>\%
  extract_features()
}
