% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install_featuretools.R
\name{install_featuretools}
\alias{install_featuretools}
\title{Install featuretools}
\usage{
install_featuretools()
}
\description{
A shortcut for installing `featuretools` with `pip`.
}
\examples{
featuretoolsR::install_featuretools()
}
