% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_primitives.R
\name{list_primitives}
\alias{list_primitives}
\title{List all available primitives.}
\usage{
list_primitives()
}
\value{
A list of all primitives available.
}
\description{
List all available primitives from `featuretools` which can be passed to \link[featuretoolsR]{dfs}.
}
\examples{
\donttest{
featuretoolsR::list_primitives()
}
}
