% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save_features.R
\name{save_features}
\alias{save_features}
\title{Save features}
\usage{
save_features(.data, filename = NA, path = NA)
}
\arguments{
\item{.data}{The tibble of features returned from `extract_features`.}
}
\description{
Used to save all or a subset of features created during `dfs`.
}
\examples{
library(magrittr)
options(stringsAsFactors = T)
set_1 <- data.frame(key = 1:100, value = sample(letters, 100, T))
set_2 <- data.frame(key = 1:100, value = sample(LETTERS, 100, T))
# Common variable: `key`

as_entityset(set_1, index = "key", entity_id = "set_1", id = "demo") \%>\%
  add_entity(entity_id = "set_2", df = set_2, index = "key") \%>\%
  add_relationship(set1 = "set_1", set2 = "set_2", idx = "key") \%>\%
  dfs(target_entity = "set_1", trans_primitives = c("and", "divide")) \%>\%
  extract_features() \%>\%
  save_features()
}
