% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy_feature_matrix.R
\name{tidy_feature_matrix}
\alias{tidy_feature_matrix}
\title{Tidy feature matrix}
\usage{
tidy_feature_matrix(.data, remove_nzv = F, nan_is_na = F)
}
\arguments{
\item{.data}{The featuretools-object returned from `dfs`.}

\item{remove_nzv}{Remove near zero variance variables created from `dfs`.}

\item{nan_is_na}{Turn all `NaN` into `NA`.}
}
\value{
A tidy data.frame.
}
\description{
Used for tidying up ('R-ify') the feature matrix after deep feature synthethis (`dfs(...)`).
}
\examples{
library(magrittr)
options(stringsAsFactors = T)
set_1 <- data.frame(key = 1:100, value = sample(letters, 100, T))
set_2 <- data.frame(key = 1:100, value = sample(LETTERS, 100, T))
# Common variable: `key`

as_entityset(set_1, index = "key", entity_id = "set_1", id = "demo") \%>\%
  add_entity(entity_id = "set_2", df = set_2, index = "key") \%>\%
  add_relationship(set1 = "set_1", set2 = "set_2", idx = "key") \%>\%
  dfs(target_entity = "set_1", trans_primitives = c("and", "divide")) \%>\%
  tidy_feature_matrix(remove_nzv = T, nan_is_na = T)
}
