% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_relationship.R
\name{add_relationship}
\alias{add_relationship}
\title{Add a relationship to an entityset}
\usage{
add_relationship(entityset, parent_set, child_set, idx)
}
\arguments{
\item{entityset}{The entityset to modify.}

\item{parent_set}{The name of the parent set.}

\item{child_set}{The name of the child set.}

\item{idx}{The variable  `parent_set` and `child_set` have in common.}
}
\value{
A modified entityset.
}
\description{
Add a relationship to an entityset.
}
\examples{
library(magrittr)
options(stringsAsFactors = TRUE)
set_1 <- data.frame(key = 1:100, value = sample(letters, 100, TRUE))
set_2 <- data.frame(key = 1:100, value = sample(LETTERS, 100, TRUE))
# Common variable: `key`

as_entityset(set_1, index = "key", entity_id = "set_1", id = "demo") \%>\%
  add_entity(entity_id = "set_2", df = set_2, index = "key") \%>\%
  add_relationship(parent_set = "set_1", child_set = "set_2", idx = "key")
}
