% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_features.R
\name{extract_features}
\alias{extract_features}
\title{Extract features}
\usage{
extract_features(.data)
}
\arguments{
\item{.data}{The featuretools-object returned from \link[featuretoolsR]{dfs}.}
}
\value{
All features created during \link[featuretoolsR]{dfs}, as a tibble.
}
\description{
This function is used to extract all features created from \link[featuretoolsR]{dfs}.
}
\examples{
\donttest{
library(magrittr)
set_1 <- data.frame(key = 1:100, value = sample(letters, 100, TRUE), stringsAsFactors = TRUE)
set_2 <- data.frame(key = 1:100, value = sample(LETTERS, 100, TRUE), stringsAsFactors = TRUE)
# Common variable: `key`

as_entityset(set_1, index = "key", entity_id = "set_1", id = "demo") \%>\%
  add_entity(entity_id = "set_2", df = set_2, index = "key") \%>\%
  add_relationship(
    parent_set = "set_1",
    child_set = "set_2",
    parent_idx = "key",
    child_idx = "key"
  ) \%>\%
  dfs(target_entity = "set_1", trans_primitives = c("and")) \%>\%
  extract_features()
}
}
