% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_features.R
\name{load_features}
\alias{load_features}
\title{Load features}
\usage{
load_features(file = NA)
}
\arguments{
\item{file}{The file containing the features.}
}
\description{
Used to load previously saved features created during `dfs`.
}
